Update to schemas 0.5.0

+++ plugins/radar-android-application-status/src/main/java/org/radarbase/monitor/application/ApplicationState.kt
@@ -32,7 +32,7 @@ class ApplicationState : BaseSourceState() {
     val cachedRecords: MutableMap<String, Long> = ConcurrentHashMap()
 
     @Synchronized
-    fun addRecordsSent(nRecords: Int) {
-        recordsSent += nRecords.toLong()
+    fun addRecordsSent(nRecords: Long) {
+        recordsSent += nRecords
     }
 }
